requirements.txt
-i https://mirrors.aliyun.com/pypi/simple/
--extra-index-url https://pypi.org/simple
hydra-core==1.3.2
numpy<2.0
einops
h5py
ftfy
regex
tqdm
timm
transformers
torchsde
torchdiffeq